Ensure carbon run payload carries bank and coverage data
